Refactor as Python Module
